+++ tests/stub.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from 'tdd-buffet/expect/jest';
 import { describe, it } from 'tdd-buffet/suite/node';
-import { ApplyProp, createStub } from '../src/mock';
+import { ApplyProp } from '../src/expectation';
+import { createStub } from '../src/mock';
 import { SingletonPendingExpectation } from '../src/pending-expectation';
 import { OneIncomingExpectationRepository } from './expectation-repository';
 import { spyExpectationFactory } from './expectations';
